+++ test/Auction/auctionTest.t.sol
@@ -8,14 +8,13 @@ import "forge-std/console.sol";
 
 import "../../src/Auction/auction.sol";
 
-contract AuctionTest is Test{
-
+contract AuctionTest is Test {
     Auction auction;
     MyNFT nftToken;
     address seller;
     address bidder1;
     address bidder2;
-    uint idToken;
+    uint256 idToken;
 
     function setUp() public {
         auction = new Auction();
@@ -32,71 +31,69 @@ contract AuctionTest is Test{
         vm.startPrank(seller);
         idToken = 123;
         nftToken.mint(seller, idToken);
-        uint prePrice = 100;
-        uint duration = 7 days;
+        uint256 prePrice = 100;
+        uint256 duration = 7 days;
         nftToken.setApprovalForAll(address(auction), true);
         auction.initAuction(duration, nftToken, idToken, prePrice);
-        assertEq(nftToken.ownerOf(idToken),address(auction));  // nft moved to auction
+        assertEq(nftToken.ownerOf(idToken), address(auction)); // nft moved to auction
         vm.stopPrank();
     }
 
     function testAddBid() public {
         vm.startPrank(bidder1);
-        uint price = 300;
+        uint256 price = 300;
         vm.deal(bidder1, price);
 
         auction.addBid(price, seller);
-        
+
         vm.stopPrank();
     }
 
     function testNotAddBid() public {
         vm.startPrank(bidder1);
-        uint price = 300;
+        uint256 price = 300;
         vm.deal(bidder1, price);
 
         vm.expectRevert();
         auction.addBid(price, address(this)); // seller is not exist
-        
+
         vm.stopPrank();
     }
 
     function testAdd2Bidds() public {
-        
         vm.startPrank(bidder1);
-        uint price = 300;
+        uint256 price = 300;
         vm.deal(bidder1, price);
 
         auction.addBid(price, seller);
-        
+
         vm.stopPrank();
-        
+
         vm.startPrank(bidder2);
         price = 500;
         vm.deal(bidder1, price);
 
         auction.addBid(price, seller);
-        
+
         vm.stopPrank();
     }
 
-     function testNotAdd2Bidds() public {
-        
+    function testNotAdd2Bidds() public {
         vm.startPrank(bidder1);
-        uint price = 300;
+        uint256 price = 300;
         vm.deal(bidder1, price);
 
         auction.addBid(price, seller);
-        
+
         vm.stopPrank();
-        
+
         vm.startPrank(bidder2);
         price = 100;
         vm.deal(bidder1, price);
 
         vm.expectRevert();
         auction.addBid(price, seller);
-        
+
         vm.stopPrank();
     }
 
@@ -113,16 +110,18 @@ contract AuctionTest is Test{
         vm.expectRevert();
         auction.removeBid(address(this));
         vm.stopPrank();
-    }  
+    }
 
-    function testEndAuction() public {         
+    function testEndAuction() public {
         testAdd2Bidds();
         vm.startPrank(seller);
-        vm.warp(block.timestamp + 7 days);        
+        vm.warp(block.timestamp + 7 days);
         auction.endAuction();
         vm.stopPrank();
     }
-    function testEndAuctionNoBidds() public {   // the seller is highest bid
+
+    function testEndAuctionNoBidds() public {
+        // the seller is highest bid
         vm.startPrank(seller);
         vm.warp(block.timestamp + 7 days);
         vm.expectRevert();
@@ -130,7 +129,8 @@ contract AuctionTest is Test{
         vm.stopPrank();
     }
 
-    function testNotEndAuction() public {  // auction not finished
+    function testNotEndAuction() public {
+        // auction not finished
         vm.startPrank(seller);
         vm.warp(block.timestamp + 4 days);
         vm.expectRevert();
@@ -138,7 +138,8 @@ contract AuctionTest is Test{
         vm.stopPrank();
     }
 
-    function testEndAuctionNotExist() public {  // user read function is not the seller
+    function testEndAuctionNotExist() public {
+        // user read function is not the seller
         vm.warp(block.timestamp + 7 days);
         vm.expectRevert();
         auction.endAuction();
